requirements.txt
torch
transformers
sentencepiece
protobuf
scipy
tqdm
edge-tts
google-genai
python-dotenv
